move test to the relevant location
